refactor: data class들 state로 이동
