+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/viewmodel/state/HomeState.kt
@@ -7,14 +7,11 @@
 
 package team.duckie.app.android.feature.ui.home.viewmodel.state
 
+import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.persistentListOf
 import team.duckie.app.android.feature.ui.home.constants.HomeStep
 import team.duckie.app.android.feature.ui.home.screen.BottomNavigationStep
-import team.duckie.app.android.feature.ui.home.screen.HomeRecommendJumbotron
-import team.duckie.app.android.feature.ui.home.screen.RecommendUserByTopic
-import team.duckie.app.android.feature.ui.home.screen.TestMaker
-import team.duckie.app.android.feature.ui.home.screen.TopicRecommendItem
 
 data class HomeState(
     val step: BottomNavigationStep = BottomNavigationStep.HomeScreen,
@@ -25,4 +22,51 @@ data class HomeState(
 
     val recommendFollowing: PersistentList<RecommendUserByTopic> = persistentListOf(),
     val recommendFollowingTest: PersistentList<TestMaker> = persistentListOf(),
-)
+) {
+    data class TestMaker(
+        val coverUrl: String,
+        val title: String,
+        val examineeNumber: Int,
+        val createAt: String,
+        val owner: User,
+    ) {
+        data class User(
+            val name: String,
+            val profile: String,
+        )
+    }
+
+    data class RecommendUserByTopic(
+        val topic: String,
+        val users: ImmutableList<RecommendUser>,
+    )
+
+    data class RecommendUser(
+        val userId: Int,
+        val profile: String,
+        val name: String,
+        val examineeNumber: Int,
+        val createAt: String,
+    )
+
+    data class HomeRecommendJumbotron(
+        val image: String,
+        val title: String,
+        val content: String,
+        val buttonContent: String,
+    )
+
+    data class TopicRecommendItem(
+        val title: String,
+        val tag: String,
+        val items: PersistentList<DuckTest>,
+    ) {
+        data class DuckTest(
+            val coverImg: String,
+            val nickname: String,
+            val title: String,
+            val examineeNumber: Int,
+            val recommendId: Int,
+        )
+    }
+}
